Use cached Intl.NumberFormat everywhere (#8244)
