+++ src/core/core.intl.js
@@ -1,0 +1,17 @@
+
+const intlCache = new Map();
+
+export function getNumberFormat(locale, options) {
+	options = options || {};
+	const cacheKey = locale + JSON.stringify(options);
+	let formatter = intlCache.get(cacheKey);
+	if (!formatter) {
+		formatter = new Intl.NumberFormat(locale, options);
+		intlCache.set(cacheKey, formatter);
+	}
+	return formatter;
+}
+
+export function formatNumber(num, locale, options) {
+	return getNumberFormat(locale, options).format(num);
+}
